requirements.txt
fastapi==0.115.11
pydantic-geojson==0.1.1
requests==2.32.3
PyJWT==2.10.1
python-dotenv==1.0.1
sqlalchemy==2.0.39
pydantic-settings
fastapi[standard]<1.0.0,>=0.114.2
geoalchemy2
psycopg2-binary
geojson-pydantic>=1.2.0
python-dateutil